src/jet-mail/requirements.txt
functions-framework==3.*
google-cloud-logging==3.10.*
mailjet-rest==1.*
flask==2.*